refactor: start simplifying directory structure. mv reused/* util/.
