+++ lib/bin/backup.js
@@ -8,7 +8,7 @@ const tmp = require('tmp');
 const archiver = require('archiver');
 const { google } = require('googleapis');
 const Problem = require('../problem');
-const Option = require('../reused/option');
+const Option = require('../util/option');
 const { connect } = require('../model/database');
 const { getOrElse } = require('../util/http');
 const { generateLocalCipherer } = require('../util/crypto');
