+++ lib/model/instance/actor.js
@@ -1,7 +1,7 @@
 const Instance = require('./instance');
 const ActeeTrait = require('../trait/actee');
 const Problem = require('../../problem');
-const { reject } = require('../../reused/promise');
+const { reject } = require('../../util/promise');
 const { withCreateTime, withUpdateTime } = require('../../util/instance');
 const { ensureArray } = require('../../util/util');
