+++ lib/model/instance/audit.js
@@ -1,5 +1,5 @@
 const Instance = require('./instance');
-const Option = require('../../reused/option');
+const Option = require('../../util/option');
 
 module.exports = Instance(({ Audit, simply }) => class {
   forCreate() { return this.with({ loggedAt: new Date() }); }
