+++ lib/model/instance/auth.js
@@ -1,5 +1,5 @@
 const Instance = require('./instance');
-const { reject } = require('../../reused/promise');
+const { reject } = require('../../util/promise');
 const Problem = require('../../problem');
 
 module.exports = Instance(({ Grant }) => class {
