+++ lib/model/instance/blob.js
@@ -1,7 +1,7 @@
 const Instance = require('./instance');
 const shasum = require('sha');
 const { readFile } = require('fs');
-const { ExplicitPromise } = require('../../reused/promise');
+const { ExplicitPromise } = require('../../util/promise');
 
 module.exports = Instance(({ all, Blob, simply }) => class {
   create() {
