+++ lib/model/instance/form.js
@@ -7,8 +7,8 @@ const ActeeTrait = require('../trait/actee');
 const Problem = require('../../problem');
 const { isBlank } = require('../../util/util');
 const { withCreateTime } = require('../../util/instance');
-const Option = require('../../reused/option');
-const { resolve, reject } = require('../../reused/promise');
+const Option = require('../../util/option');
+const { resolve, reject } = require('../../util/promise');
 
 const formFields = [ 'id', 'xmlFormId', 'xml', 'version', 'hash', 'name', 'acteeId', 'createdAt', 'updatedAt', 'deletedAt' ];
 Object.freeze(formFields);
