+++ lib/model/instance/grant.js
@@ -1,8 +1,8 @@
 const Instance = require('./instance');
 const { ensureArray } = require('../../util/util');
 const { withCreateTime } = require('../../util/instance');
-const { resolve } = require('../../reused/promise');
-const Option = require('../../reused/option');
+const { resolve } = require('../../util/promise');
+const Option = require('../../util/option');
 
 module.exports = Instance(({ all, grants }) => class {
   forCreate() { return withCreateTime(this); }
