+++ lib/model/instance/session.js
@@ -1,6 +1,6 @@
 const { withCreateTime } = require('../../util/instance');
 const Instance = require('./instance');
-const Option = require('../../reused/option');
+const Option = require('../../util/option');
 const { generateToken } = require('../../util/crypto');
