+++ lib/model/instance/submission.js
@@ -6,8 +6,8 @@ const Instance = require('./instance');
 const Problem = require('../../problem');
 const { isBlank } = require('../../util/util');
 const { withCreateTime } = require('../../util/instance');
-const Option = require('../../reused/option');
-const { resolve, reject } = require('../../reused/promise');
+const Option = require('../../util/option');
+const { resolve, reject } = require('../../util/promise');
 
 
 const submissionFields = [ 'id', 'formId', 'instanceId', 'xml', 'submitter', 'createdAt', 'updatedAt', 'deletedAt' ];
