+++ lib/model/query/field-keys.js
@@ -1,5 +1,5 @@
 const { merge } = require('ramda');
-const Option = require('../../reused/option');
+const Option = require('../../util/option');
 const { fieldsForJoin, joinRowToInstance, maybeFirst } = require('../../util/db');
 
 module.exports = {
