+++ lib/model/query/future-query.js
@@ -1,6 +1,6 @@
 const { postgresErrorToProblem } = require('../../util/db');
 const { merge, identity, compose } = require('ramda');
-const { maybePoint, reject } = require('../../reused/promise');
+const { maybePoint, reject } = require('../../util/promise');
 
 
 // The general premise behind the entire query system is as follows:
