+++ lib/model/query/grants.js
@@ -1,6 +1,6 @@
 
 const { rowsToInstances } = require('../../util/db');
-const Option = require('../../reused/option');
+const Option = require('../../util/option');
 
 // shared fragment which takes the target actor id and any relevant actor system
 // ids and recursively resolves all group memberships into a seq of actor ids.
