+++ lib/model/query/submissions.js
@@ -1,6 +1,6 @@
 const Problem = require('../../problem');
 const { fieldsForJoin, joinRowToInstance, rowsToInstances, maybeFirst, translateProblem } = require('../../util/db');
-const { resolve } = require('../../reused/promise');
+const { resolve } = require('../../util/promise');
 
 module.exports = {
   getById: (formId, instanceId, extended) => ({ submissions }) =>
