+++ lib/resources/config.js
@@ -3,8 +3,8 @@ const config = require('config');
 const { isBlank, printPairs } = require('../util/util');
 const { endpoint, getOrNotFound, getOrElse, success } = require('../util/http');
 const Problem = require('../problem');
-const Option = require('../reused/option');
-const { ExplicitPromise, reject } = require('../reused/promise');
+const Option = require('../util/option');
+const { ExplicitPromise, reject } = require('../util/promise');
 const { generateKeypair } = require('../util/crypto');
 
 // TODO: not here.
