+++ lib/resources/submissions.js
@@ -3,8 +3,8 @@ const { always } = require('ramda');
 const sanitize = require('sanitize-filename');
 const { openRosaEndpoint, endpoint, getOrNotFound, getOrReject, rejectIf } = require('../util/http');
 const { openRosaMessage } = require('../outbound/openrosa');
-const { resolve, reject } = require('../reused/promise');
-const Option = require('../reused/option');
+const { resolve, reject } = require('../util/promise');
+const Option = require('../util/option');
 const Problem = require('../problem');
 const { streamJoinedCsvs } = require('../data/csv');
 const { streamAttachments } = require('../data/attachments');
