+++ lib/resources/users.js
@@ -1,7 +1,7 @@
 const { endpoint, getOrNotFound, success, isTrue } = require('../util/http');
 // const { verifyPassword } = require('../util/crypto');
 const Problem = require('../problem');
-const { resolve, reject } = require('../reused/promise');
+const { resolve, reject } = require('../util/promise');
 
 module.exports = (service, { Actee, User, mail }) => {
