+++ lib/util/crypto.js
@@ -2,7 +2,7 @@ const bcrypt = require('bcrypt');
 const { generatePrivateKey, createPublicKey, createPrivateKey } = require('ursa');
 const { randomBytes, pbkdf2, createCipheriv, createDecipheriv } = require('crypto');
 const { promisify } = require('util');
-const { ExplicitPromise, resolve } = require('../reused/promise');
+const { ExplicitPromise, resolve } = require('../util/promise');
 const { isBlank } = require('./util');
