+++ lib/util/db.js
@@ -1,7 +1,7 @@
 const { inspect } = require('util');
-const { reject } = require('../reused/promise');
+const { reject } = require('../util/promise');
 const Problem = require('../problem');
-const Option = require('../reused/option');
+const Option = require('../util/option');
 
 
 ////////////////////////////////////////////////////////////////////////////////
