+++ lib/util/http.js
@@ -5,8 +5,8 @@ const { DateTime } = require('luxon');
 const { isBlank } = require('./util');
 const { openRosaMessage } = require('../outbound/openrosa');
 const Problem = require('../problem');
-const Option = require('../reused/option');
-const { reject } = require('../reused/promise');
+const Option = require('../util/option');
+const { reject } = require('../util/promise');
 
 const isTrue = (x) => (!isBlank(x) && (x.toLowerCase() === 'true'));
