+++ lib/util/option.js
