+++ lib/util/promise.js
